backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
pandas==2.2.2
numpy==2.0.2
scikit-learn==1.5.2
statsmodels==0.14.3
celery==5.4.0
redis==5.0.7
psycopg2-binary==2.9.9
requests==2.32.3
joblib==1.4.2
openpyxl==3.1.5
